requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
tqdm>=4.66.0
networkx>=3.2.0
scikit-learn>=1.3.0
hydra-core>=1.3.2 